chore(test-utils): categorize unit tests
