+++ packages/test-utils/tests/helpers/__tests__/patterns.test.ts
@@ -4,9 +4,9 @@ import {
   createCircuitBreakerHarness,
   createErrorScenarioBuilder,
   simulateRetry,
-} from '../src/patterns';
+} from '../../../src/patterns';
 
-import { advanceTimersAndFlush } from '../src/helpers';
+import { advanceTimersAndFlush } from '../../../src/helpers';
 
 describe('patterns', () => {
   it('simulates circuit breaker lifecycle', async () => {
